+++ src/rules/color-no-invalid-hex/index.js
@@ -20,6 +20,10 @@ export default function (actual) {
       const declString = decl.toString()
 
       styleSearch({ source: declString, target: "#" }, match => {
+        // If there's not a colon, comma, or whitespace character before, we'll assume this is
+        // not intended to be a hex color, but is instead something like the
+        // has in a url() argument
+        if (!/[:,\s]/.test(declString[match.startIndex - 1])) { return }
 
         const hexMatch = /^#[0-9A-Za-z]+/.exec(declString.substr(match.startIndex))
         if (!hexMatch) { return }